feat: integrate feature development repo into main codebase
- Migrated components, pages, and utilities developed in standalone feature repo
- Resolved conflicts while preserving changes from both codebases
- Verified route components export valid React components (page/layout)
- Updated import paths and ensured compatibility with Next.js App Router
- Fixed runtime errors due to missing or incorrect exports
