+++ frontend/src/app/devices/[id]/page.tsx
@@ -3,8 +3,26 @@ import React, { useState } from "react";
 import { useParams, useRouter } from "next/navigation";
 import { useSearchParams } from "next/navigation";
 import { FiHome, FiMonitor, FiLink, FiBarChart2 } from "react-icons/fi";
-import ThemeToggle from "@/app/theme-toggle";
-import ConnectionDetails from "@/components/ConnectionDetails";
+import { ThemeToggle } from "@/app/theme-toggle";
+import { ConnectionDetails } from "@/app/components/ConnectionDetails";
+/**
+ * Renders the DevicePage component, showing detailed information about a specific device.
+ *
+ * Includes a sidebar for navigation and tabbed sections for various device data,
+ * such as overview, connection details, and connection charts.
+ *
+ * @remarks
+ * - Designed for client-side rendering only, as it relies on `useParams` and `useSearchParams`.
+ * - Uses a responsive layout that adjusts based on sidebar visibility.
+ * - Handles active tab state and sidebar toggle logic.
+ * - Icons from `react-icons` visually represent each tab.
+ * - Includes a Home button for quick navigation.
+ *
+ * @returns The rendered device detail page.
+ *
+ * @see {@link ConnectionDetails} for displaying device interface details.
+ * @see {@link ThemeToggle} for the theme switching functionality.
+ */
 
 interface TabItem {
   label: string;
