+++ frontend/src/app/types/graphql/GetDeviceInterfaces.ts
