+++ frontend/src/app/types/graphql/GetZoneDevices.ts
