+++ frontend/src/app/utils/stringUtils.ts
